migrateファイルの修正

+++ db/migrate/20250522072452_drop_posts_table.rb
@@ -1,6 +1,6 @@
 class DropPostsTable < ActiveRecord::Migration[8.0]
   def up
-    drop_table :posts
+    drop_table :posts if table_exists?(:posts)
   end
 
   def down
